Add config template to repo.
The config template is used by Docker - it sets most of the configuration via envrioment variables.
As the container is built from the git repo, it needs a copy of the configuration it it. I've added
it under a name that won't be used by errbot locally, but allows docker to find it :).